refactor: split ec2 into ec2_images & ec2_instances
