+++ tests/test_ec2_instances.py
@@ -6,18 +6,7 @@ from moto import mock_ec2
 from moto.ec2 import ec2_backends
 from moto.ec2.models import AMIS
 
-from aec.command.ec2 import (
-    delete_image,
-    describe,
-    describe_images,
-    launch,
-    logs,
-    modify,
-    share_image,
-    start,
-    stop,
-    terminate,
-)
+from aec.command.ec2_instances import describe, launch, logs, modify, start, stop, terminate
 
 
 @pytest.fixture
@@ -157,29 +146,6 @@ def describe_instance0(region_name, instance_id):
     return instances["Reservations"][0]["Instances"][0]
 
 
-def test_describe_images(mock_aws_config):
-    # describe images defined by moto
-    # see https://github.com/spulec/moto/blob/master/moto/ec2/resources/amis.json
-    canonical_account_id = "099720109477"
-    mock_aws_config["describe_images_owners"] = canonical_account_id
-    images = describe_images(config=mock_aws_config)
-
-    assert len(images) == 2
-    assert images[0]["Name"] == "ubuntu/images/hvm-ssd/ubuntu-trusty-14.04-amd64-server-20170727"
-    assert images[1]["Name"] == "ubuntu/images/hvm-ssd/ubuntu-xenial-16.04-amd64-server-20170721"
-
-
-def test_describe_images_name_match(mock_aws_config):
-    # describe images defined by moto
-    # see https://github.com/spulec/moto/blob/master/moto/ec2/resources/amis.json
-    canonical_account_id = "099720109477"
-    mock_aws_config["describe_images_owners"] = canonical_account_id
-    images = describe_images(config=mock_aws_config, name_match="*trusty*")
-
-    assert len(images) == 1
-    assert images[0]["Name"] == "ubuntu/images/hvm-ssd/ubuntu-trusty-14.04-amd64-server-20170727"
-
-
 def test_stop_start(mock_aws_config):
     launch(mock_aws_config, "alice", AMIS[0]["ami_id"])
 
@@ -198,10 +164,6 @@ def test_modify(mock_aws_config):
     assert instances[0]["Type"] == "c5.2xlarge"
 
 
-def test_delete_image(mock_aws_config):
-    delete_image(mock_aws_config, AMIS[0]["ami_id"])
-
-
 def test_terminate(mock_aws_config):
     launch(mock_aws_config, "alice", AMIS[0]["ami_id"])
 
@@ -212,10 +174,6 @@ def test_logs(mock_aws_config):
     launch(mock_aws_config, "alice", AMIS[0]["ami_id"])
 
     logs(config=mock_aws_config, name="alice")
-
-
-def test_share_image(mock_aws_config):
-    share_image(mock_aws_config, AMIS[0]["ami_id"], "123456789012")
 
 
 @pytest.mark.skip(
